feat/flow users transfers

+++ api/database/migrations/2021_04_02_181845_create_wallets_table.php
@@ -16,7 +16,7 @@ class CreateWalletsTable extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
             $table->string('owner')->unique();
-            $table->integer('value');
+            $table->float('value', 8, 4);
             $table->timestamps();
         });
     }
